Buddy_Backend/requirements.txt
# If this was the combined result you want:
fastapi
uvicorn
sentence-transformers
faiss-cpu
openai
